requirements.txt
xmind-sdk-python>=1.2.0
markdown>=3.4.0
beautifulsoup4>=4.11.0